refactor: icon 이름 변경 및 index 추가(#15)
