+++ src/components/Icons/SmileFaceIcon.tsx
