+++ src/components/Icons/index.ts
@@ -21,3 +21,4 @@ export { default as ReplayIcon } from './ReplayIcon'
 export { default as SettingsIcon } from './SettingsIcon'
 export { default as TrashIcon } from './TrashIcon'
 export { default as UserIcon } from './UserIcon'
+export { default as SmileFaceIcon } from './SmileFaceIcon'
